refactor: Use AbstractExtraction as a parent for all extraction classes
