+++ app/supplejack/extraction/enrichment_extraction.rb
@@ -1,9 +1,7 @@
 # frozen_string_literal: true
 
 module Extraction
-  class EnrichmentExtraction
-    attr_accessor :document
-
+  class EnrichmentExtraction < AbstractExtraction
     def initialize(extraction_definition, record, page, extraction_folder = nil)
       @extraction_definition = extraction_definition
       @record = record
@@ -11,24 +9,9 @@ module Extraction
       @extraction_folder = extraction_folder
     end
 
-    def extract
-      Sidekiq.logger.info 'Fetching records from the Enrichment Source'
-      @document = Extraction::Request.new(url:, params:, headers:).get
-    end
-
-    def save
-      raise ArgumentError, 'extraction_folder was not provided in #new' unless @extraction_folder.present?
-      raise '#extract must be called before #save DocumentExtraction' unless @document.present?
-
-      @document.save(file_path)
-    end
-
-    def extract_and_save
-      extract
-      save
-    end
-
     private
+
+    def params; end
 
     # rubocop:disable Lint/UnusedBlockArgument
     def url
@@ -41,15 +24,6 @@ module Extraction
       name_str = @extraction_definition.name.parameterize(separator: '_')
       page_str = format('%09d', @page)[-9..]
       "#{@extraction_folder}/#{name_str}__#{@record['id']}__#{page_str}.json"
-    end
-
-    def params; end
-
-    def headers
-      {
-        'Content-Type' => 'application/json',
-        'User-Agent' => 'Supplejack Harvester v2.0'
-      }
     end
   end
 end
