+++ app/supplejack/extraction/record_extraction.rb
@@ -1,16 +1,9 @@
 module Extraction
-  class RecordExtraction
-    attr_accessor :document
-
+  class RecordExtraction < AbstractExtraction
     def initialize(extraction_definition, page)
       @extraction_definition = extraction_definition
       @api_source            = extraction_definition.destination
       @page = page
-    end
-
-    def extract
-      Sidekiq.logger.info 'Fetching records from the API'
-      @document = Extraction::Request.new(url:, params:, headers:).get
     end
 
     private
@@ -28,13 +21,6 @@ module Extraction
           page: @page
         },
         api_key: @api_source.api_key
-      }
-    end
-
-    def headers
-      {
-        'Content-Type' => 'application/json',
-        'User-Agent' => 'Supplejack Harvester v2.0'
       }
     end
   end
